fix flappers
still fugly
